More Refactoring : Remove data abstraction
Previously, we were trying to abstract data into reusable structs. However, this started to get a bit unwieldy, as we had to have data structs for API response, manifest response, and then the abstracted data. We could end up with 3 data types representing parts of one item (say an activity).

So, now, ApiInterface and ManifestInterface just return the data objects for the underlying data, and the program can then combine them as needed.