Added an ActivityStatsContainer, which contains a Vec<Activity> and provides access to aggregated stats for all of the activities.
Right now, this is specific to Activity, but I may create a trait for
items that have game stats, so we can work with different data.

Note, there are some instances where standing can missing from api
activity responses. I am pretty sure this is a bug, so instead of
setting the property to Option<Standing>, i set it to default of
Standing::Unknown.

Also, i havent implimented draws yet for standing, as I need to confirm
the value returned from the API.

dcliah is basically working now, i.e. it takes arguments, and gets the
correct data in the format we need. now just need to spit it out for the
user.