Mini refactor for how we pass around member ids. Now, instead of passing member_id, character_id, platform to each function and letting it find the data in the db, we pass around the appropriate db index for the member / char.
This reduces the complexity of some of the sql / db calls, and also
reduces the number of calls we need to make.

Also move inserting member / charter to the front of sync, to ensure
there is valid data before we start to sync with the API.

Ran fmt format with some new controls too.