store all modes in db for each activity. (needed for searching on submodes).
Added crucible module / classes, to abstract crucible results.

currently structured around single player results, and not an activity,
with multiple players (i.e. a player results contains an activity). Will
refactor this if we ever decide to storate all player data for an
activity, and not just the specified player.