Display win % when getting kills with a specific weapon
Now, when displaying weapon info in dcliah, we also display the win percentage when getting a kill with each weapon.

Note, this does not include activities (wins / loss) when using the weapon, but registering no kills, only activities in which the user got a kill in the game.

Not sure how useful this is, so for right now, this should be considered a beta feature.