Fixed issue where wrong data was being retrieved for member
If data was mnissing on a member from API, we would go back to server to get updated data. However, we were matching based on cross save, and not on member_id, which meant the data wasnt actually being updated (we were adding a new member).

Fixed by matching on member_id now.